style(steps): rename to kebab-case
