+++ week-1-ts-ecosystem/w01e05-my-own-cli/src/steps/ask-for-generate-data.mts
