+++ week-1-ts-ecosystem/w01e05-my-own-cli/src/steps/get-product-type.mts
